Some thoughts on how this could work
Created a new table in the database with same structure and messages as before, but with additional parent_id column. Default value is 0, which means it's a first message in a conversation. All the reply messages get their parent_id set to the id of the first message.
There are some TODOs in the code, also one that's not mentioned: the functions should not return anything if the input params are bad (this is related to Joonas' suggestion).